Updates gallery functionality
