+++ js/gallery.js
@@ -1,41 +1,21 @@
-import {isEscEvent} from './util.js';
-import {fullsize, fullsizeCancel, renderFullsize} from './render-fullsize-picture.js';
+import {openFullsize} from './fullsize-picture.js';
 
-const body = document.querySelector('body');
+const addPreviewClickHandlers = (data) => {
+  const previewSection = document.querySelector('.pictures');
 
-const closeFullsize = () => {
-  fullsize.classList.add('hidden');
-  body.classList.remove('modal-open');
-};
-
-const keyEscHandler = (evt) => {
-  if (isEscEvent(evt)) {
-    evt.preventDefault();
-    closeFullsize();
-  }
-};
-
-const openFullsize = () => {
-  fullsize.classList.remove('hidden');
-  body.classList.add('modal-open');
-  document.addEventListener('keydown', keyEscHandler);
-};
-
-const previewClickHandler = (data) => {
-  const previews = document.querySelectorAll('.picture');
-  previews.forEach((item, idx) => {
-    item.addEventListener('click', (evt) => {
+  const previewClickHandler = (evt) => {
+    if (evt.target.closest('.picture')) {
       evt.preventDefault();
-      openFullsize();
-      renderFullsize(data[idx]);
-    });
-  });
+      const onePreview = evt.target.closest('.picture');
+      const onePreviewId = +onePreview.getAttribute('data-id');
+
+      openFullsize(data.find(({id}) => id === onePreviewId));
+    }
+  };
+
+  previewSection.addEventListener('click', previewClickHandler);
 };
 
-fullsizeCancel.addEventListener('click', () => {
-  closeFullsize();
-  document.removeEventListener('keydown', keyEscHandler);
-});
 
-export {previewClickHandler};
+export {addPreviewClickHandlers};
