+++ js/main.js
@@ -1,12 +1,10 @@
 import {createSimilarPosts} from './data.js';
 import {renderAllPictures} from './render-pictures.js';
-import {previewClickHandler} from './gallery.js';
+import {addPreviewClickHandlers} from './gallery.js';
+
 
 const POSTS_COUNT = 25;
 const similarPictures = createSimilarPosts(POSTS_COUNT);
 
 renderAllPictures(similarPictures);
-
-previewClickHandler(similarPictures);
-
-
+addPreviewClickHandlers(similarPictures);
