+++ js/render-pictures.js
@@ -1,12 +1,13 @@
 const pictureList = document.querySelector('.pictures');
 const pictureTemplate = document.querySelector('#picture').content.querySelector('.picture');
 
-const renderOnePicture = ({url, likes, comments}) => {
+const renderOnePicture = ({id, url, likes, comments}) => {
   const pictureElement = pictureTemplate.cloneNode(true);
   const pictureImg = pictureElement.querySelector('.picture__img');
   const pictureLikes = pictureElement.querySelector('.picture__likes');
   const pictureComments = pictureElement.querySelector('.picture__comments');
 
+  pictureElement.setAttribute('data-id', id);
   pictureImg.src = url;
   pictureLikes.textContent = likes;
   pictureComments.textContent = comments.length;
